% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_response}
\alias{gh_response}
\title{gh_response}
\usage{
gh_response(
  issue,
  comment,
  owner,
  repo,
  figure = paste0("issue_figures/", issue, ".png"),
  branch = "main"
)
}
\arguments{
\item{issue}{github issue number}

\item{comment}{text message to comment on issue}

\item{owner}{github user name}

\item{repo}{github repository}

\item{figure}{figure path in the working repository (default: `issue_figures/{issue}.png`)}

\item{branch}{github branch to commit (default: "main")}
}
\description{
submit a comment with an attached figure
}
\author{
Chenhao Li
}
